src/test/java/algorithms/leetcode/ContainerWithMostWaterTest.java
package algorithms.leetcode;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class ContainerWithMostWaterTest {

    private ContainerWithMostWater containerWithMostWater;

    @BeforeEach
    void setUp() {
        containerWithMostWater = new ContainerWithMostWater();
    }

    @ParameterizedTest(name = "Params: input=`{0}`, expected=`{1}`")
    @MethodSource("params")
    void shouldCalculateMostWaterContainerArea(int[] input, int expected) {
        //when
        int result = containerWithMostWater.maxArea(input);
        //then
        assertThat(result).isEqualTo(expected);
    }

    static Stream<Arguments> params() {
        return Stream.of(
                arguments(new int[]{1, 8, 6, 2, 5, 4, 8, 3, 7}, 49),
                arguments(new int[]{1, 2, 1}, 2),
                arguments(new int[]{1, 1}, 1)
        );
    }
}